requirements.txt
aiosmtplib==2.0.0
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
black==22.12.0
blinker==1.5
boto3==1.26.59
botocore==1.29.59
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==38.0.4
databases==0.6.1
decouple==0.0.7
dnspython==2.2.1
docopt==0.6.2
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.88.0
fastapi-mail==1.2.4
fastapi-utils==0.2.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
ormar==0.12.0
passlib==1.7.4
pathspec==0.10.3
pipreqs==0.4.11
platformdirs==2.6.2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
starlette==0.22.0
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
yarg==0.1.9
